mv tests, crate docs
